+++ lib/services/apiAuthService.js
@@ -1,7 +1,7 @@
 const fetch = require('isomorphic-fetch');
 const logger = require('../logger')
 const config = require('../config');
-const uriFragSplitter = require('../uriFragSplitter');
+const uriFragSplitter = require('../helpers/uriFragSplitter');
 const varifySession = require('../middleware/verifySession');
 const serviceUri = config.get('API_GATEWAY_HOST');
 const clientId = config.get('API_AUTH_CLIENT_ID');
